requirements.txt
aiogram==2.25
aiohttp==3.8.6
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==22.2.0
Babel==2.9.1
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
colorama==0.4.6
crashtest==0.4.1
cryptography==39.0.2
Cython @ file:///buildstream-build
distlib==0.3.6
dulwich==0.21.3
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.9.0
frozenlist==1.4.1
greenlet==3.0.3
html5lib==1.1
icecream==2.1.3
idna==3.4
importlib-metadata==6.0.0
installer==0.6.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
magic-filter==1.0.12
Mako @ file:///buildstream-build
Markdown @ file:///buildstream-build
MarkupSafe @ file:///buildstream-build
meson==0.63.3
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
packaging==23.0
pandas==2.2.2
pexpect==4.8.0
pipenv==2023.2.18
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
ptyprocess==0.7.0
pycparser==2.21
Pygments @ file:///buildstream-build
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
pytz==2024.1
rapidfuzz==2.13.7
requests==2.28.2
requests-toolbelt==0.10.1
SecretStorage==3.3.3
shellingham==1.5.0.post1
six @ file:///buildstream-build
SQLAlchemy==2.0.29
tomli==2.0.1
tomlkit==0.11.6
trove-classifiers==2023.3.9
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.15
virtualenv==20.20.0
virtualenv-clone==0.5.7
webencodings==0.5.1
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
yarl==1.9.4
zipp==3.15.0
